requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
email-validator==2.1.1
WTForms-SQLAlchemy==0.4.1
